작업내용 : 머지 후 충돌 처리 완료

+++ LL2_Client_Win_Source/ItemPacketHandler.h
@@ -12,7 +12,7 @@ public:
 	static void HandleUseItemResult(const ParsedPacket& pkt);
 	static void HandlePickUpItem(const ParsedPacket& pkt);
 	static void SendUseItem(InventoryItemInfo* inventoryitemInfo);
-	static void SendPickupDropItem(int dropId);
+	static void SendPickupDropItem(int64_t dropId);
 	
 private:
 };
